refactor: AvailablePoint - point 패키지로 이동
